requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.8
scipy==1.11.3
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
Pillow==10.1.0 